requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
Authlib
pydantic[email]
bcrypt==4.0.1
